+++ atlasdb-dbkvs-tests/src/test/java/com/palantir/atlasdb/keyvalue/dbkvs/DbkvsPostgresKeyValueServiceTest.java
@@ -15,8 +15,20 @@
  */
 package com.palantir.atlasdb.keyvalue.dbkvs;
 
+import static org.junit.Assert.assertEquals;
+
+import java.nio.charset.StandardCharsets;
+import java.util.Map;
+
+import org.junit.Test;
+
+import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.ImmutableSet;
+import com.palantir.atlasdb.AtlasDbConstants;
+import com.palantir.atlasdb.keyvalue.api.Cell;
 import com.palantir.atlasdb.keyvalue.api.KeyValueService;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
+import com.palantir.atlasdb.keyvalue.api.Value;
 import com.palantir.atlasdb.keyvalue.dbkvs.impl.ConnectionManagerAwareDbKvs;
 import com.palantir.atlasdb.keyvalue.impl.AbstractKeyValueServiceTest;
 
@@ -30,5 +42,31 @@ public class DbkvsPostgresKeyValueServiceTest extends AbstractKeyValueServiceTes
             }
         }
         return kvs;
+    }
+
+    @Test
+    public void handlesLongTablenamesWithoutMergingThem() {
+        // Create two empty tables with names only unique well past the name limit
+        TableReference longTablename1 = TableReference.createFromFullyQualifiedName(
+                "atlas.ThisIsAVeryLongTableNameThatWillExceedEvenTheLongerPostgresLimit");
+        TableReference longTablename2 = TableReference.createFromFullyQualifiedName(
+                "atlas.ThisIsAVeryLongTableNameThatWillExceedEvenTheLongerPostgresLimitAsWell");
+        keyValueService.createTables(ImmutableMap.of(
+                longTablename1, AtlasDbConstants.GENERIC_TABLE_METADATA,
+                longTablename2, AtlasDbConstants.GENERIC_TABLE_METADATA));
+        keyValueService.truncateTables(ImmutableSet.of(longTablename1, longTablename2));
+
+        // Put some data in table 1
+        Cell key = Cell.create("foo".getBytes(StandardCharsets.UTF_8), "bar".getBytes(StandardCharsets.UTF_8));
+        byte[] value = "baz".getBytes(StandardCharsets.UTF_8);
+        keyValueService.put(longTablename1, ImmutableMap.of(key, value), 0L);
+
+        // Try and retrieve the data we put in table 1 from table 2
+        Map<Cell, Value> cellValueMap = keyValueService.get(longTablename2, ImmutableMap.of(key, Long.MAX_VALUE));
+
+        // If the data is present in table 2, the tables have merged
+        assertEquals(cellValueMap, ImmutableMap.of());
+
+        keyValueService.dropTables(ImmutableSet.of(longTablename1, longTablename2));
     }
 }