+++ atlasdb-dbkvs/src/main/java/com/palantir/atlasdb/keyvalue/dbkvs/OracleTableNameMapper.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2016 Palantir Technologies
+ * Copyright 2017 Palantir Technologies
  *
  * Licensed under the BSD-3 License (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,80 +15,16 @@
  */
 package com.palantir.atlasdb.keyvalue.dbkvs;
 
-import com.google.common.base.Preconditions;
-import com.palantir.atlasdb.AtlasDbConstants;
-import com.palantir.atlasdb.keyvalue.api.Namespace;
-import com.palantir.atlasdb.keyvalue.api.TableReference;
-import com.palantir.atlasdb.keyvalue.dbkvs.impl.ConnectionSupplier;
-import com.palantir.atlasdb.keyvalue.dbkvs.impl.DbKvs;
-import com.palantir.nexus.db.sql.AgnosticResultSet;
+public class OracleTableNameMapper extends TableNameMapper {
 
-public class OracleTableNameMapper {
-    public static final int ORACLE_MAX_TABLE_NAME_LENGTH = 30;
-    public static final int SUFFIX_NUMBER_LENGTH = 6;
-    public static final int MAX_NAMESPACE_LENGTH = 2;
-    private static final int PREFIXED_TABLE_NAME_LENGTH = ORACLE_MAX_TABLE_NAME_LENGTH - SUFFIX_NUMBER_LENGTH;
-
-    public String getShortPrefixedTableName(
-            ConnectionSupplier connectionSupplier,
-            String tablePrefix,
-            TableReference tableRef) {
-        Preconditions.checkState(tablePrefix.length() <= AtlasDbConstants.MAX_TABLE_PREFIX_LENGTH,
-                "The tablePrefix can be at most %s characters long", AtlasDbConstants.MAX_TABLE_PREFIX_LENGTH);
-
-        TableReference shortenedNamespaceTableRef = truncateNamespace(tableRef);
-
-        String prefixedTableName = tablePrefix + DbKvs.internalTableName(shortenedNamespaceTableRef);
-        String truncatedTableName = truncate(prefixedTableName, PREFIXED_TABLE_NAME_LENGTH);
-
-        String fullTableName = tablePrefix + DbKvs.internalTableName(tableRef);
-        return truncatedTableName + getTableNumberSuffix(connectionSupplier, fullTableName, truncatedTableName);
+    @Override
+    public int getMaxTableNameLength() {
+        return 30;
     }
 
-    private TableReference truncateNamespace(TableReference tableRef) {
-        if (tableRef.getNamespace().isEmptyNamespace()) {
-            return tableRef;
-        }
-        String namespace = tableRef.getNamespace().getName();
-        namespace = truncate(namespace, MAX_NAMESPACE_LENGTH);
-        return TableReference.create(Namespace.create(namespace), tableRef.getTablename());
+    @Override
+    public int getShortenedNamespaceLength() {
+        return 2;
     }
 
-    private String truncate(String name, int length) {
-        return name.substring(0, Math.min(name.length(), length));
-    }
-
-    private String getTableNumberSuffix(
-            ConnectionSupplier connectionSupplier,
-            String fullTableName,
-            String truncatedTableName) {
-        int tableSuffixNumber = getNextTableNumber(connectionSupplier, truncatedTableName);
-        if (tableSuffixNumber >= 100_000) {
-            throw new IllegalArgumentException(
-                    "Cannot create any more tables with name starting with " + truncatedTableName
-                    + ". 100,000 tables might have already been created. Please rename the table." + fullTableName);
-        }
-        return "_" + String.format("%05d", tableSuffixNumber);
-    }
-
-    private int getNextTableNumber(ConnectionSupplier connectionSupplier, String truncatedTableName) {
-        AgnosticResultSet results = connectionSupplier.get().selectResultSetUnregisteredQuery(
-                "SELECT short_table_name "
-                + "FROM " + AtlasDbConstants.ORACLE_NAME_MAPPING_TABLE
-                + " WHERE LOWER(short_table_name) LIKE LOWER(?||'\\______%') ESCAPE '\\'"
-                + " ORDER BY short_table_name DESC", truncatedTableName);
-        return getTableNumberFromTableNames(truncatedTableName, results);
-    }
-
-    private int getTableNumberFromTableNames(String truncatedTableName, AgnosticResultSet results) {
-        for (int i = 0; i < results.size(); i++) {
-            String shortName = results.get(i).getString("short_table_name");
-            try {
-                return Integer.parseInt(shortName.substring(truncatedTableName.length() + 1)) + 1;
-            } catch (NumberFormatException e) {
-                //Table in different format - Do nothing;
-            }
-        }
-        return 0;
-    }
 }