+++ atlasdb-dbkvs/src/main/java/com/palantir/atlasdb/keyvalue/dbkvs/impl/oracle/OracleWriteTable.java
@@ -17,17 +17,17 @@ package com.palantir.atlasdb.keyvalue.dbkvs.impl.oracle;
 
 import com.palantir.atlasdb.keyvalue.api.TableReference;
 import com.palantir.atlasdb.keyvalue.dbkvs.DdlConfig;
-import com.palantir.atlasdb.keyvalue.dbkvs.OracleTableNameGetter;
+import com.palantir.atlasdb.keyvalue.dbkvs.TableNameGetter;
 import com.palantir.atlasdb.keyvalue.dbkvs.impl.AbstractDbWriteTable;
 import com.palantir.atlasdb.keyvalue.dbkvs.impl.ConnectionSupplier;
-import com.palantir.atlasdb.keyvalue.dbkvs.impl.OraclePrefixedTableNames;
+import com.palantir.atlasdb.keyvalue.dbkvs.impl.MappedTableNames;
 
 public class OracleWriteTable extends AbstractDbWriteTable {
     public OracleWriteTable(
             DdlConfig config,
             ConnectionSupplier conns,
-            OracleTableNameGetter oracleTableNameGetter,
+            TableNameGetter tableNameGetter,
             TableReference tableRef) {
-        super(config, conns, tableRef, new OraclePrefixedTableNames(config, conns, oracleTableNameGetter));
+        super(config, conns, tableRef, new MappedTableNames(config, conns, tableNameGetter));
     }
 }