+++ docs/source/release_notes/release-notes.rst
@@ -42,6 +42,9 @@ develop
     *    - Type
          - Change
 
+    *    - Type
+         - Change
+
     *    - |improved|
          - Increase default Cassandra pool size from minimum of 20 and maximum of 5x the minimum (100 if minimum not modified) connections to minimum of 30 and maximum of 100 connections.
            This allows for better handling of bursts of requests that would otherwise require creating many new connections to Cassandra from the clients.
@@ -71,6 +74,10 @@ develop
          - All SnapshotTransaction get methods are now safe for tables declared with SweepStrategy.THOROUGH.
            Previously, a validation check was omitted for ``getRowsColumnRange``, ``getRowsIgnoringLocalWrites``, and ``getIgnoringLocalWrites``, which in very rare cases could have resulted in deleted values being returned by a long-running read transaction.
            (`Pull Request <https://github.com/palantir/atlasdb/pull/1421>`__)
+
+    *    - |fixed|
+         - Fixed an issue where Postgres KVS mishandled tablenames of >63 characters.
+           (`Pull Request <https://github.com/palantir/atlasdb/pull/1441>`__)
 
 .. <<<<------------------------------------------------------------------------------------------------------------->>>>